feat: Landmarks for simple matches
Add a landmark region for elements that match the query and that don't
already have a role, aria-labelledby or aria-label attribute. Also,
clean up the region when the highlight is removed.

The `aria-roledescription` attribute is used to provide a description of
"Highlight"—the region's name is just a unique number.

This doesn't work on elements that already have those attributes; they
will need to be wrapped in a landmark region.